feat: rewrite files module, use convict for config validation and refactor everything

+++ src/modules/sessions/sessions.controller.ts
@@ -1,17 +1,17 @@
-import { plainToClass } from "class-transformer";
-
 import { ApiExcludeEndpoint } from "@nestjs/swagger";
-import { Controller, Delete, Get, Param, Session, UseGuards } from "@nestjs/common";
+import { Controller, Delete, Get, Param, Req, UseGuards } from "@nestjs/common";
+
+import { plainToClass } from "class-transformer";
 
 import { SessionsService } from "./sessions.service";
 
-import { SessionDto } from "./dto/session.dto";
-
-import { ISessionData } from "./interfaces/session-data.interface";
-
 import { CurrentUser } from "@/decorators/current-user.decorator";
 
+import { SessionDto } from "@/dto/session.dto";
+
 import { AuthGuard } from "@/guards/auth.guard";
+
+import { Request } from "@/interfaces/request.interface";
 
 @Controller("sessions")
 @UseGuards(AuthGuard)
@@ -20,35 +20,38 @@ export class SessionsController {
 
   @ApiExcludeEndpoint()
   @Get()
-  async getSessions(
+  async getAll(
     @CurrentUser("id") uid: string,
-    @Session() currentSession: ISessionData
+    @Req() req: Request
   ): Promise<SessionDto[]> {
     const sessions = await this.sessions.find({ "session.uid": uid });
 
     return sessions.map(({ session }) =>
       plainToClass(SessionDto, {
         ...session,
-        isCurrent: session.identifier === currentSession.identifier
+        isCurrent: session.id === req.session.id
       })
     );
   }
 
   @ApiExcludeEndpoint()
-  @Delete("/:id/revoke")
-  revoke(@CurrentUser("id") uid: string, @Param("id") id: string): Promise<void> {
-    return this.sessions.deleteOne({
-      "session.identifier": id,
+  @Delete()
+  revokeAll(
+    @CurrentUser("id") uid: string,
+    @Req() req: Request
+  ): Promise<void> {
+    return this.sessions.deleteMany({
+      id: { $ne: req.session.id },
       "session.uid": uid
     });
   }
 
   @ApiExcludeEndpoint()
-  @Delete("all")
-  revokeAll(@CurrentUser("id") uid: string, @Session() session: ISessionData): Promise<void> {
-    return this.sessions.delete({
-      "session.identifier": { $ne: session.identifier },
-      "session.uid": uid
-    });
+  @Delete(":id/revoke")
+  revoke(
+    @CurrentUser("id") uid: string,
+    @Param("id") id: string
+  ): Promise<void> {
+    return this.sessions.deleteOne({ id, "session.uid": uid });
   }
 }
